Remove SdkApi and merge functionality into EmbraceInternalInterface
