+++ embrace-android-sdk/api/embrace-android-sdk.api
@@ -40,7 +40,6 @@ public final class io/embrace/android/embracesdk/Embrace : io/embrace/android/em
 	public fun getInternalInterface ()Lio/embrace/android/embracesdk/internal/EmbraceInternalInterface;
 	public fun getLastRunEndState ()Lio/embrace/android/embracesdk/Embrace$LastRunEndState;
 	public fun getReactNativeInternalInterface ()Lio/embrace/android/embracesdk/ReactNativeInternalInterface;
-	public fun getSdkApi ()Lio/embrace/android/embracesdk/internal/SdkApi;
 	public fun getSessionProperties ()Ljava/util/Map;
 	public fun getTraceIdHeader ()Ljava/lang/String;
 	public fun getUnityInternalInterface ()Lio/embrace/android/embracesdk/UnityInternalInterface;
@@ -178,9 +177,8 @@ public final class io/embrace/android/embracesdk/WebViewClientSwazzledHooks {
 public abstract interface annotation class io/embrace/android/embracesdk/annotation/StartupActivity : java/lang/annotation/Annotation {
 }
 
-public abstract interface class io/embrace/android/embracesdk/internal/SdkApi {
-	public abstract fun getSdkCurrentTime ()J
 public abstract interface class io/embrace/android/embracesdk/internal/EmbraceInternalInterface {
+	public abstract fun getSdkCurrentTime ()J
 	public abstract fun isNetworkSpanForwardingEnabled ()Z
 	public abstract fun logComposeTap (Landroid/util/Pair;Ljava/lang/String;)V
 	public abstract fun logError (Ljava/lang/String;Ljava/util/Map;Ljava/lang/String;Z)V
