--- embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/SdkApiImpl.kt
@@ -1,7 +1,0 @@
-package io.embrace.android.embracesdk.internal
-
-import io.embrace.android.embracesdk.clock.Clock
-
-internal class SdkApiImpl(private val clock: Clock) : SdkApi {
-    override fun getSdkCurrentTime(): Long = clock.now()
-}
